avoid restarting the readline loop
Since apparently haskeline doesn't provide us with
a way to replace the completion function, the
old code restarted the haskeline InputT monad.

We take another approach: we use a stateful completion
function. The design of Haskeline lets us use any monad,
but as we are already in IO we just use IORefs.

Later we can move to a solution without mutation - e.g.
StateT or EnvT should we wish. Moreover lenses work
fine with composite monadic state so we will get rid of the 5
state variables we pass around between loop and initLoop
as well.

Note that now we also save the history to cubicaltt.history.txt

I avoided something like ~/.local/share/cubicaltt/history or
~/.cubicaltt as I'd like compatibility with Windows and haskeline
apparently doesn't provide it out of the box for the filename.