split Main.initLoop into compileFile, handleErrors and resumeLoop
While overall code size is not any smaller, the control flow
is much easier to grasp now:

initLoop flags f completionRef
  = compileFile f >>= handleErrors >>= resumeLoop flags f completionRef

It turns out that various error handling options did the same thing:
they output the message, prepared new (names, tenv) and restarted
the loop.

So now

- the golden path (without any errors) is shown very clearly in
  compileFile
- the error handling is in one function with just one parameter - it
  turns out it doesn't need the 5 parameters we pass around
- compileFile doesn't need them either, it only ever requires file name

wrapTypeChecker wrapper is nasty - but now this nastiness doesn't affect
us when we look at the main control path, and it can be reduced by
changing the interface of runDeclss which turned out to be
not convenient at all.

IO lifting can be refactored now too, as the full InputM monad
transformer is only used in the resumeLoop part